// src/shared/schemas.ts
export const ConsumerRelationshipSchema = z.object({
  producerSlug: z.string(),
  producerName: z.string(),
  bruceApiUrl: z.string(),
  mappingFile: z.string(),
});

export const ConsumerIndexSchema = z.object({
  relationships: z.array(ConsumerRelationshipSchema),
});
export type ConsumerIndex = z.infer<typeof ConsumerIndexSchema>;